+++ lib/lf/refinement/hybrid2d_refinement_pattern.cc
@@ -45,7 +45,7 @@ std::ostream &operator<<(std::ostream &o, const RefPat &refpat) {
 }
 
 // Implementation for RefinemeentPattern
-
+// NOLINTNEXTLINE(readability-function-cognitive-complexity)
 lf::base::size_type Hybrid2DRefinementPattern::NumChildren(
     lf::base::dim_t codim) const {
   LF_VERIFY_MSG(codim <= ref_el_.Dimension(),
@@ -268,6 +268,7 @@ lf::base::size_type Hybrid2DRefinementPattern::NumChildren(
 }
 
 std::vector<Eigen::Matrix<int, Eigen::Dynamic, Eigen::Dynamic>>
+// NOLINTNEXTLINE(readability-function-cognitive-complexity)
 Hybrid2DRefinementPattern::ChildPolygons(lf::base::dim_t codim) const {
   LF_VERIFY_MSG(codim <= ref_el_.Dimension(),
                 "Illegal codim = " << codim << " for " << ref_el_.ToString());
@@ -276,9 +277,9 @@ Hybrid2DRefinementPattern::ChildPolygons(lf::base::dim_t codim) const {
   std::vector<Eigen::Matrix<int, Eigen::Dynamic, Eigen::Dynamic>> child_poly{};
 
   // Lattice point coordinates (lattice_const_ should be a multiple of 6)
-  const unsigned int lt_half = lattice_const_ / 2;
-  const unsigned int lt_third = lattice_const_ / 3;
-  const unsigned int lt_one = lattice_const_;
+  const int lt_half = base::narrow<int>(lattice_const_ / 2);
+  const int lt_third = base::narrow<int>(lattice_const_ / 3);
+  const int lt_one = base::narrow<int>(lattice_const_);
   // Depending on the type of cell do something different
   switch (ref_el_) {
     case lf::base::RefEl::kPoint(): {
